check eval sheet and testet my mains. some stuff to do

+++ cpp_07/ex01/iter.hpp
@@ -4,9 +4,11 @@
 # include <iostream>
 
 template <typename T, typename F>
-void iter(T *t, int length, F f)
+// TODO: WE changed the second arg to size_t!
+void iter(T *t, size_t length, F f)
 {
-    for (int i = 0; i < length; i++)
+	// TODO: HERE AS WELL: WE changed i to size_t!
+    for (size_t i = 0; i < length; i++)
         f(t[i]);
 }
